+++ Example/AEViewModelExample/GithubTVC.swift
@@ -7,7 +7,7 @@
 import AEViewModel
 import SafariServices
 
-final class GithubTVMC: TableViewModelController {
+final class GithubTVC: TableViewController {
 
     // MARK: Properties
 
@@ -52,7 +52,7 @@ final class GithubTVMC: TableViewModelController {
         return .customNib(GithubRepoCell.self)
     }
 
-    override func action(for cell: TableViewModelCell, at indexPath: IndexPath, sender: Any) {
+    override func action(for cell: TableCell, at indexPath: IndexPath, sender: Any) {
         if let repo = dataSource.viewModel(at: indexPath) as? Repo, let url = URL(string: repo.url) {
             let browser = SFSafariViewController(url: url)
             browser.title = title