+++ Sources/Cell.swift
@@ -6,22 +6,22 @@
 
 import UIKit
 
-public protocol ViewModelCellDelegate: class {
-    func callback(from cell: ViewModelCell, sender: Any)
+public protocol CellDelegate: class {
+    func callback(from cell: Cell, sender: Any)
 }
 
-public protocol ViewModelCell: class {
-    var delegate: ViewModelCellDelegate? { get set }
+public protocol Cell: class {
+    var delegate: CellDelegate? { get set }
     var userInfo: [AnyHashable : Any] { get set }
 
     func configure()
     func reset()
     func update(with item: Item)
 
-    func callback(_ sender: Any)
+    func callback(userInfo: [AnyHashable: Any]?, sender: Any)
 }
 
-public extension ViewModelCell {
+public extension Cell {
     static var reuseIdentifier: String {
         return String(describing: self)
     }